% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_rpe_to_pct.R
\name{create_map_rep_rpe_to_pct}
\alias{create_map_rep_rpe_to_pct}
\title{Create a mapping from reps and rpe to pct_1rm}
\usage{
create_map_rep_rpe_to_pct(pct_lookup_tab)
}
\arguments{
\item{pct_lookup_tab}{A lookup table that contains columns reps, rpe and pct_1rm.}
}
\value{
A function taking arguments reps and rpe, returning a numeric vector.
}
\description{
Create a mapping from reps and rpe to pct_1rm
}
\examples{
resulting_function <- create_map_rep_rpe_to_pct(pct_lookup_rts)
}
