% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e1rm.R
\name{e1rm}
\alias{e1rm}
\title{Estimated one rep max}
\usage{
e1rm(weight, reps, rpe, pct1rm_func)
}
\arguments{
\item{weight}{Numeric vector of weights, one entry per set.}

\item{reps}{Numeric vector of reps, one entry per set.}

\item{rpe}{Numeric vector of RPEs, one entry per set.}

\item{pct1rm_func}{A function which takes \code{reps},\code{rpe} and returns an estimated percentage of one rep max.}
}
\value{
A numeric vector of estimated one rep max values for each set.
}
\description{
Estimated one rep max
}
