% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_sets.R
\name{hard_sets}
\alias{hard_sets}
\title{Number of hard sets}
\usage{
hard_sets(rpe, cutoff = 7)
}
\arguments{
\item{rpe}{A numeric vector of set RPE values.}

\item{cutoff}{RPE cutoff for what qualifies as a hard set.  The boundary is included.}
}
\value{
A numeric value counting the number of hard sets.
}
\description{
Number of hard sets
}
\examples{
hard_sets(c(6, 7, 10, 9, NA))
}
