% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pct_lookup_brzycki}
\alias{pct_lookup_brzycki}
\title{Percentage lookup from Brzycki.}
\format{
\subsection{\code{pct_lookup_brzycki}}{

A data frame with 135 rows and 3 columns:
\describe{
\item{reps}{Number of reps performed}
\item{rpe}{Rating of perceived exertion}
\item{pct_1rm}{Percentage of one-repetition maximum associated with \code{reps} at \code{rpe}.}
}
}
}
\source{
{Derived from \code{rpe_tab_brzycki}.}
}
\usage{
pct_lookup_brzycki
}
\description{
Percentage lookup from Brzycki.
}
\keyword{datasets}
