% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pct_lookup_helms}
\alias{pct_lookup_helms}
\title{Percentage lookup from Helms and colleagues.}
\format{
\subsection{\code{pct_lookup_helms}}{

A data frame with 90 rows and 3 columns:
\describe{
\item{reps}{Number of reps performed}
\item{rpe}{Rating of perceived exertion}
\item{pct_1rm}{Percentage of one-repetition maximum associated with \code{reps} at \code{rpe}.}
}
}
}
\source{
{Derived from \code{rpe_tab_helms}.}
}
\usage{
pct_lookup_helms
}
\description{
Percentage lookup from Helms and colleagues.
}
\keyword{datasets}
