% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pct_lookup_rts}
\alias{pct_lookup_rts}
\title{Percentage lookup from reactive training systems (RTS)}
\format{
\subsection{\code{pct_lookup_rts}}{

A data frame with 165 rows and 5 columns:
\describe{
\item{reps}{Number of reps performed}
\item{rpe}{Rating of perceived exertion}
\item{next_rpe_down}{Next rating of perceived exertion in the table provided}
\item{pct_1rm}{Percentage of one-repetition maximum associated with \code{reps} at \code{rpe}.}
\item{next_pct_1rm_down}{Percentage of one-repetition maximum associated with \code{reps} at \code{next_rpe_down}.}
}
}
}
\source{
{Derived from \code{rpe_tab_rts}.}
}
\usage{
pct_lookup_rts
}
\description{
Percentage lookup from reactive training systems (RTS)
}
\keyword{datasets}
