% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_rpe_to_pct.R
\name{rep_rpe_to_pct_rts}
\alias{rep_rpe_to_pct_rts}
\title{Mapping from reps and rpe to pct_1rm using Reactive Training System's table.}
\usage{
rep_rpe_to_pct_rts(reps, rpe)
}
\arguments{
\item{reps}{A vector of repetitions performed in a set.}

\item{rpe}{A vector of ratings of perceived exertion.}
}
\value{
A numeric vector listing the percentage of one 1 rep max.
}
\description{
Mapping from reps and rpe to pct_1rm using Reactive Training System's table.
}
\details{
\code{reps} must be length 1 or the same length as \code{rpe}.  RPE must be length 1 or the same length as \code{reps}.
}
\examples{
resulting_function <- create_map_rep_rpe_to_pct(pct_lookup_rts)
}
