% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpe_tab_brzycki}
\alias{rpe_tab_brzycki}
\title{RPE table from Brzycki.}
\format{
\subsection{\code{rpe_tab_brzycki}}{

A dataframe with 9 rows and 16 columns.
\describe{
\item{RPE}{The rating of perceived exertion}
\item{1}{Estimated percentage of one rep max doing 1 rep at the RPE given by RPE.}
\item{2}{Estimated percentage of one rep max doing 2 reps at the RPE given by RPE.}'
\item{...}{And so on...}
}
}
}
\source{
{ISBN 978-1935628132, "A practical approach to strength training".  This is a rep-max resource and we extended to RPE using the reps in reserve definition for RPE (see \code{rep_tab_helms} for a paper link that explains this.)}
}
\usage{
rpe_tab_brzycki
}
\description{
RPE table from Brzycki.
}
\keyword{datasets}
