% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpe_tab_helms}
\alias{rpe_tab_helms}
\title{RPE table Helms and colleagues.}
\format{
\subsection{\code{rpe_tab_helms}}{

A dataframe with 10 rows and 11 columns.
\describe{
\item{RPE}{The rating of perceived exertion}
\item{1}{Estimated percentage of one rep max doing 1 rep at the RPE given by RPE.}
\item{2}{Estimated percentage of one rep max doing 2 reps at the RPE given by RPE.}'
\item{...}{And so on...}
}
}
}
\source{
{https://doi.org/10.1519/SSC.0000000000000218.  Based on earlier work by Zourdos and Helms}
}
\usage{
rpe_tab_helms
}
\description{
RPE table Helms and colleagues.
}
\keyword{datasets}
