% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpe_tab_rts}
\alias{rpe_tab_rts}
\title{RPE table RTS}
\format{
\subsection{\code{rpe_tab_rts}}{

A dataframe with 11 rows and 16 columns.
\describe{
\item{RPE}{The rating of perceived exertion}
\item{1}{Estimated percentage of one rep max doing 1 rep at the RPE given by RPE.}
\item{2}{Estimated percentage of one rep max doing 2 reps at the RPE given by RPE.}'
\item{...}{And so on...}
}
}
}
\source{
{Pulled from the TRAC system on Reactive Training Systems in Dec 2025}
}
\usage{
rpe_tab_rts
}
\description{
RPE table RTS
}
\keyword{datasets}
