% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpe_tab_to_lookup.R
\name{rpe_tab_to_lookup}
\alias{rpe_tab_to_lookup}
\title{RPE table to lookup table}
\usage{
rpe_tab_to_lookup(rpe_tab)
}
\arguments{
\item{rpe_tab}{A tibble with columns RPE, 1, 2, 3, ... containing the percentage of one rep max associated with doing x reps at RPE rating of perceived exertion.}
}
\value{
A tibble with columns rpe, reps, pct_1rm.
}
\description{
RPE table to lookup table
}
\examples{
rpe_tab_to_lookup(rpe_tab_rts)
}
